home page finished

+++ app/controllers/votes_controller.rb
@@ -12,9 +12,15 @@ class VotesController < ApplicationController
 
   def create
     @question = Question.find(params[:question_id])
+    if current_user == nil
+      flash[:error] = "You should sign in first"
+      redirect_to question_path(@question) 
+      return 
+    end
     @new_vote = Vote.new
-    @new_vote.update_attributes(:user_id => current_user.id, :question_id => @question.id) 
-    if @new_vote.save
+    if Vote.where(:user_id => current_user.id, :question_id => @question.id).first == nil 
+      @new_vote.update_attributes(:user_id => current_user.id, :question_id => @question.id) 
+      @new_vote.save
       @question.update_attributes(:vote_count => @question.vote_count+1)
       if @question.save
         redirect_to question_path(@question)
@@ -36,6 +42,11 @@ class VotesController < ApplicationController
 
   def destroy
     @question = Question.find(params[:question_id])
+    if current_user == nil
+      flash[:error] = "You should sign in first"
+      redirect_to question_path(@question) 
+      return 
+    end
     @vote = Vote.where(question_id: @question.id, user_id:current_user.id)
     if @vote.length != 0 and @vote.destroy_all
       @question.update_attributes(:vote_count => @question.vote_count-1)
